server/requirements.txt
Flask==3.1.3
flask-cors==6.0.1
PyJWT==2.10.1
PyMySQL==1.1.2
cryptography==46.0.3
pypdf==6.4.1
python-docx==1.2.0
langchain==0.3.28
langchain-chroma==0.2.6
langchain-openai==0.3.35
chromadb==1.5.7
python-dotenv==1.2.1
